Add preliminary log bus concept
The log bus provides a queue for log sources to dump their data into.
Sinks ought to "connect" to the other end of the bus and push the data
towards persistent storage or apply filtering and so forth.

Implement a simple bufferbus which just consists of a single buffer
which is written to from start to end. Said buffer is allocated as a
memory mapped file. This bus does not synchronize writing and reading.
It is therefore useful for single threaded work where log entries can be
periodically flushed down the sinks by the thread itself.

Implement the ringbus which is powered by multiple bipartite buffers all
residing in a memory mapped file. The bipartite buffers use lock free
synchronization and are layed out to avoid false sharing between
concurrent writers and the singular allowed reader.