feat!: Replace generic attributes in records
After reevaluating the record format, it became clear that there are
only the file and line attributes which actually need to be attached to
each record. In virtually all other cases it is way more efficient to
tie attributes to the logger instance and just stream changes to the
logger's attribute set.